+++ bootique-cayenne-ehcache/src/test/java/io/bootique/cayenne/ehcache/persistent/Table2.java
@@ -1,0 +1,11 @@
+package io.bootique.cayenne.ehcache.persistent;
+
+import io.bootique.cayenne.ehcache.persistent.auto._Table2;
+import org.apache.cayenne.lifecycle.cache.CacheGroups;
+
+@CacheGroups("g3")
+public class Table2 extends _Table2 {
+
+    private static final long serialVersionUID = 1L; 
+
+}